% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.fit.R
\docType{data}
\name{ergm.fit}
\alias{ergm.fit}
\title{Output from an Exponential Random Graph Model fit on the Mesa High Network (Hunter et. al, 2008).}
\format{An object of class \code{"ergm"}}
\usage{
data(ergm.fit)
}
\description{
This is the output of a call to the ergm function fit with with an edge term,
  sex, grade, and race homophily, GWESP, and alternating k-stars.  Fit according to the vignette "mesaHighVignette".
}
\references{
Hunter et al. (2008) ergm: A package to fit, simulate and diagnose exponential-family models for networks. Journal of Statistical Software 24(3).
}
\keyword{datasets}
