% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fergm.fit.R
\docType{data}
\name{fergm.fit}
\alias{fergm.fit}
\title{Output from a Frailty Exponential Random Graph Model fit on the Mesa High Network (Hunter et. al, 2008)}
\format{A list containing an object of class \code{"stan.fit"}}
\usage{
data(fergm.fit)
}
\description{
This is the output of a call to the ergm function fit with with an edge term,
  sex, grade, and race homophily, GWESP, and alternating k-stars.  Fit according to "mesaHighVignette".
}
\references{
Hunter et al. (2008) ergm: A package to fit, simulate and diagnose exponential-family models for networks. Journal of Statistical Software 24(3).
}
\keyword{datasets}
