% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesa.R
\docType{data}
\name{mesa}
\alias{mesa}
\title{Mesa High Network (Hunter et. al, 2008).}
\format{An object of class \code{"network"}}
\usage{
data(mesa)
}
\description{
This is a network included in Hunter et al. (2008) that represents a simulated in-school friendship network.
}
\references{
Hunter et al. (2008) ergm: A package to fit, simulate and diagnose exponential-family models for networks. Journal of Statistical Software 24(3).
}
\keyword{datasets}
